requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
pyyaml
python-dotenv
openai>=1.13.3
pdfminer.six
python-docx